chore: Optimizers refactored.
